# swagger.yaml
...
openapi: 3.0.3
info:
  title: Clinic-AI API
  version: 1.0.0
  description: "AI-powered clinical intake system. This spec is being filled module-by-module."
servers:
  - url: http://localhost:8000
paths:
...
  /workflow/walk-in/create-visit:
    post:
      tags: [workflow]
      summary: Create walk-in visit
      description: Create a walk-in visit for patients without intake.
      operationId: create_walk_in_visit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWalkInVisitRequestSchema'
      responses:
        '201':
          description: Walk-in visit created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_CreateWalkInVisitResponseSchema'
        '400': {description: Validation error, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '500': {description: Internal server error, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
  /workflow/visit/{visit_id}/available-steps:
    get:
      tags: [workflow]
      summary: Get available workflow steps
      description: Get available workflow steps for a visit based on its workflow type and status.
      operationId: get_available_workflow_steps
      parameters:
        - name: visit_id
          in: path
          required: true
          schema: {type: string}
      responses:
        '200': {description: Available steps, content: {application/json: {schema: {$ref: '#/components/schemas/ApiResponse_Dict'}}}}
        '404': {description: Visit not found, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '500': {description: Internal server error, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
  /workflow/visits/walk-in:
    get:
      tags: [workflow]
      summary: List walk-in visits
      description: List walk-in visits with pagination.
      operationId: list_walk_in_visits
      parameters:
        - name: limit
          in: query
          required: false
          schema: {type: integer, default: 100}
        - name: offset
          in: query
          required: false
          schema: {type: integer, default: 0}
      responses:
        '200': {description: Walk-in visits listed, content: {application/json: {schema: {$ref: '#/components/schemas/ApiResponse_Dict'}}}}
        '500': {description: Internal server error, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
...
components:
  schemas:
    ErrorResponse:
      type: object
      required: [success, error, message, timestamp, request_id]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: INTERNAL_SERVER_ERROR
        message:
          type: string
          example: An unexpected internal server error occurred.
        details:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
          format: uuid
...
    ApiResponse_Dict:
      type: object
      required: [success, message, timestamp, request_id, data]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: OK
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
          format: uuid
        data:
          type: object
          additionalProperties: true
...
    CreateWalkInVisitRequestSchema:
      type: object
      required: [name, mobile]
      properties:
        name: {type: string, description: Patient name}
        mobile: {type: string, description: Patient mobile number}
        age: {type: integer, nullable: true, description: Patient age}
        gender: {type: string, nullable: true, description: Patient gender}
    CreateWalkInVisitResponseSchema:
      type: object
      required: [patient_id, visit_id, workflow_type, status, message]
      properties:
        patient_id: {type: string}
        visit_id: {type: string}
        workflow_type: {type: string}
        status: {type: string}
        message: {type: string}
    ApiResponse_CreateWalkInVisitResponseSchema:
      type: object
      required: [success, message, timestamp, request_id]
      properties:
        success: {type: boolean, example: true}
        message: {type: string, example: 'Walk-in visit created successfully'}
        timestamp: {type: string, format: date-time}
        request_id: {type: string, format: uuid}
        data:
          $ref: '#/components/schemas/CreateWalkInVisitResponseSchema'
...